backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
gunicorn
packaging
pillow
psycopg2-binary
PyJWT
python-dotenv
sqlparse
tzdata
whitenoise
